@@ -26,11 +26,7 @@ import com.facebook.drawee.backends.pipeline.Fresco;
 import java.lang.ref.WeakReference;
 import java.util.ArrayList;
 
-import eu.dkaratzas.android.inapp.update.Constants;
-import eu.dkaratzas.android.inapp.update.InAppUpdateManager;
-import eu.dkaratzas.android.inapp.update.InAppUpdateStatus;
-
-public class EntryActivity extends BaseActivity implements InAppUpdateManager.InAppUpdateHandler {
+public class EntryActivity extends BaseActivity  {
 
     private static final int REQ_CODE_VERSION_UPDATE = 530;
     private static final String TAG = "StickerEntryActivity";
@@ -55,26 +51,12 @@ public class EntryActivity extends BaseActivity implements InAppUpdateManager.In
         // AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES);
 
         // For toolbar
-        Toolbar toolbar = findViewById(R.id.app_toolbar);
-        setSupportActionBar(toolbar);
+        //Toolbar toolbar = findViewById(R.id.app_toolbar);
+        //setSupportActionBar(toolbar);
 
         progressBar = findViewById(R.id.entry_activity_progress);
         loadListAsyncTask = new LoadListAsyncTask(this);
         loadListAsyncTask.execute();
-
-        initInAppUpdate();
-    }
-
-    private void initInAppUpdate() {
-        InAppUpdateManager inAppUpdateManager = InAppUpdateManager.Builder(this, REQ_CODE_VERSION_UPDATE)
-                .resumeUpdates(true) // Resume the update, if the update was stalled. Default is true
-                //.mode(Constants.UpdateMode.FLEXIBLE)
-                .mode(Constants.UpdateMode.IMMEDIATE)
-                .snackBarMessage("An update has just been downloaded.")
-                .snackBarAction("RESTART")
-                .handler(this);
-
-        inAppUpdateManager.checkForAppUpdate();
     }
 
     private void showStickerPack(ArrayList<StickerPack> stickerPackList) {
@@ -110,16 +92,6 @@ public class EntryActivity extends BaseActivity implements InAppUpdateManager.In
         }
     }
 
-    @Override
-    public void onInAppUpdateError(int code, Throwable error) {
-        Log.d(TAG, "code: " + code, error);
-    }
-
-    @Override
-    public void onInAppUpdateStatus(InAppUpdateStatus status) {
-
-    }
-
     static class LoadListAsyncTask extends AsyncTask<Void, Void, Pair<String, ArrayList<StickerPack>>> {
         private final WeakReference<EntryActivity> contextWeakReference;
 
@@ -134,7 +106,7 @@ public class EntryActivity extends BaseActivity implements InAppUpdateManager.In
                 final Context context = contextWeakReference.get();
                 if (context != null) {
                     stickerPackList = StickerPackLoader.fetchStickerPacks(context);
-                    if (stickerPackList.size() == 0) {
+                    if (stickerPackList.isEmpty()) {
                         return new Pair<>("could not find any packs", null);
                     }
                     for (StickerPack stickerPack : stickerPackList) {
