@@ -13,7 +13,7 @@ import android.os.Parcelable;
 
 import java.util.List;
 
-class StickerPack implements Parcelable {
+public class StickerPack implements Parcelable {
     final String identifier;
     final String name;
     final String publisher;
